vcfapi/requirements.txt
amqp==2.6.1
aniso8601==7.0.0
arrow==1.2.2
asgiref==3.5.0
BaseHash==3.0.10
billiard==3.6.4.0
blessed==1.19.1
cachetools==5.0.0
celery==4.4.7
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.9
django-cors-headers==3.11.0
django-filter==21.1
django-otp==1.1.3
django-picklefield==3.1
django-q==1.3.9
django-rest-passwordreset==1.2.1
django-storages==1.12.3
django-templated-mail==1.1.1
djangorestframework==3.13.1
drf-access-policy==1.1.0
drf-writable-nested==0.6.3
drf-yasg==1.20.0
Faker==14.2.0
gcloud==0.18.3
google-api-core==2.7.1
google-api-python-client==2.28.0
google-auth==2.6.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.3
google-cloud-firestore==2.3.4
google-cloud-secret-manager==2.8.0
google-cloud-storage==2.2.1
google-cloud-tasks==2.7.1
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
googlemaps==4.5.3
graphene==2.1.9
graphene-django==2.15.0
graphene-django-optimizer==0.8.0
graphql-core==2.3.2
graphql-relay==2.0.1
grpc-google-iam-v1==0.12.4
grpcio==1.48.1
grpcio-status==1.48.1
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
kombu==4.6.11
libcst==0.4.7
libnum==1.7.1
MarkupSafe==2.1.0
mypy-extensions==0.4.3
oauth2client==4.1.3
packaging==21.3
Pillow==9.0.1
promise==2.3
proto-plus==1.22.1
protobuf==3.20.0
psycopg2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.15.0
PyJWT==2.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
redis==3.5.3
requests==2.27.1
requests-toolbelt==0.9.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
typing-inspect==0.8.0
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.8
vine==1.3.0
wcwidth==0.2.5
whitenoise==6.0.0
vobject
boto3
opentelemetry-sdk
opentelemetry-api
opentelemetry-instrumentation-django
uptrace